requirements.txt
<<<<<<< HEAD
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
pydantic==2.9.0
google-generativeai==0.8.0
duckduckgo-search==6.3.0
=======
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
pydantic==2.9.0
google-generativeai==0.8.0
duckduckgo-search==6.3.0
>>>>>>> e3bdfa0dc1e232c7a1a90430320437a724ae975b
